requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.0
python-dotenv==1.0.0
APScheduler==3.10.4
cryptography==41.0.3
python-dateutil==2.8.2
WTForms==3.0.1
Flask-WTF==1.1.1 